# src/ApiHost/Controllers/ToDoListsController.cs
using Base.Domain.Result;
using Microsoft.AspNetCore.Mvc;
using ToDoLists.Contracts;
using Wolverine;

namespace ApiHost.Controllers;

/// <summary>
/// Request model for adding a new todo item to a list.
/// </summary>
/// <param name="Title">The title of the new todo item.</param>
public sealed record AddToDoRequest(string Title);

/// <summary>
/// REST API controller for todo list management operations.
/// </summary>
[ApiController]
[Route("api/v1/todo-lists")]
[Produces("application/json")]
public class ToDoListsController : ControllerBase
{
    private readonly IMessageBus _messageBus;

    public ToDoListsController(IMessageBus messageBus)
    {
        _messageBus = messageBus;
    }

    /// <summary>
    /// Creates a new todo list in the system.
    /// </summary>
    [HttpPost]
    [ProducesResponseType(typeof(CreateToDoListResult), StatusCodes.Status201Created)]
    [ProducesResponseType(typeof(ValidationProblemDetails), StatusCodes.Status400BadRequest)]
    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
    public async Task<IActionResult> CreateToDoList([FromBody] CreateToDoListCommand command)
    {
        Result<CreateToDoListResult> result = await _messageBus.InvokeAsync<Result<CreateToDoListResult>>(command);

        if (result.IsSuccess)
        {
            return CreatedAtAction(
                nameof(GetToDoList),
                new { id = result.Value.ToDoListId },
                result.Value);
        }

        return HandleError(result.Error);
    }

    /// <summary>
    /// Retrieves a todo list by its unique identifier.
    /// </summary>
    [HttpGet("{id}")]
    [ProducesResponseType(typeof(ToDoListDetailDto), StatusCodes.Status200OK)]
    [ProducesResponseType(typeof(ValidationProblemDetails), StatusCodes.Status400BadRequest)]
    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status403Forbidden)]
    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
    public async Task<IActionResult> GetToDoList(string id, [FromQuery] string userId)
    {
        var query = new GetToDoListQuery(id, userId);
        Result<ToDoListDetailDto> result = await _messageBus.InvokeAsync<Result<ToDoListDetailDto>>(query);

        if (result.IsSuccess)
        {
            return Ok(result.Value);
        }

        return HandleError(result.Error);
    }

    /// <summary>
    /// Adds a new todo item to an existing todo list.
    /// </summary>
    [HttpPost("{listId}/todos")]
    [ProducesResponseType(typeof(AddToDoResult), StatusCodes.Status201Created)]
    [ProducesResponseType(typeof(ValidationProblemDetails), StatusCodes.Status400BadRequest)]
    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
    [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
    public async Task<IActionResult> AddToDoToList(string listId, [FromBody] AddToDoRequest request)
    {
        var command = new AddToDoCommand(listId, request.Title);
        Result<AddToDoResult> result = await _messageBus.InvokeAsync<Result<AddToDoResult>>(command);

        if (result.IsSuccess)
        {
            return CreatedAtAction(
                nameof(GetToDoList),
                new { id = listId },
                result.Value);
        }

        return HandleError(result.Error);
    }

    private IActionResult HandleError(Error error)
    {
        return error.Type switch
        {
            ErrorType.Validation => CreateValidationProblem(error),
            ErrorType.NotFound => NotFound(CreateProblemDetails("Resource not found", error)),
            ErrorType.Forbidden => StatusCode(StatusCodes.Status403Forbidden, CreateProblemDetails("Forbidden", error)),
            _ => Problem(
                title: "An error occurred while processing the request",
                detail: error.Description,
                statusCode: StatusCodes.Status500InternalServerError)
        };
    }

    private IActionResult CreateValidationProblem(Error error)
    {
        ModelState.AddModelError(string.Empty, error.Description);
        return ValidationProblem();
    }

    private ProblemDetails CreateProblemDetails(string title, Error error)
    {
        return new ProblemDetails
        {
            Title = title,
            Detail = error.Description,
            Status = GetStatusCodeForErrorType(error.Type)
        };
    }

    private static int GetStatusCodeForErrorType(ErrorType errorType)
    {
        return errorType switch
        {
            ErrorType.NotFound => StatusCodes.Status404NotFound,
            ErrorType.Validation => StatusCodes.Status400BadRequest,
            ErrorType.Forbidden => StatusCodes.Status403Forbidden,
            _ => StatusCodes.Status500InternalServerError
        };
    }
}